Rename SecurityContextStore to securityContextStore
