+++ src/Sulu/Bundle/SecurityBundle/Resources/js/containers/Permissions/Permissions.js
@@ -5,8 +5,8 @@ import {observer} from 'mobx-react';
 import {Loader} from 'sulu-admin-bundle/components';
 import {ResourceMultiSelect} from 'sulu-admin-bundle/containers';
 import {userStore} from 'sulu-admin-bundle/stores';
-import securityContextStore from '../../stores/SecurityContextStore';
-import type {SecurityContextGroups, SecurityContexts} from '../../stores/SecurityContextStore/types';
+import securityContextStore from '../../stores/securityContextStore';
+import type {SecurityContextGroups, SecurityContexts} from '../../stores/securityContextStore/types';
 import type {ContextPermission} from './types';
 import permissionsStyle from './permissions.scss';
 import PermissionMatrix from './PermissionMatrix';
