--- src/Sulu/Bundle/SecurityBundle/Resources/js/stores/SecurityContextStore/index.js
@@ -1,4 +1,0 @@
-// @flow
-import SecurityContextStore from './SecurityContextStore';
-
-export default SecurityContextStore;
