+++ src/Sulu/Bundle/SecurityBundle/Resources/js/stores/securityContextStore/index.js
@@ -1,0 +1,4 @@
+// @flow
+import securityContextStore from './securityContextStore';
+
+export default securityContextStore;
