+++ src/Sulu/Bundle/SecurityBundle/Resources/js/stores/securityContextStore/securityContextStore.js
