+++ src/Sulu/Bundle/SecurityBundle/Resources/js/stores/securityContextStore/types.js
